backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
boto3==1.29.7
requests==2.31.0
python-decouple==3.8
firebase-admin==6.2.0
geopy==2.4.0
django-filter==23.3
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
